Refactor Flocker operations into separate file
